requirements/base.txt
# Base
pytz==2018.7
python-slugify==1.2.6
Pillow==7.1.0
psycopg2==2.8.6 --no-binary psycopg2

# Django
django==3.1

# Environment
django-environ==0.4.5

# Django REST Framework
djangorestframework==3.12.1
django-filter==2.4.0
django-cors-headers >= 3.0.2

# JWT
pyjwt==1.7.1

# Passwords security
argon2-cffi==20.1.0

# Static files
whitenoise>=5.0.0

# Celery
redis>=3.2.0
django-redis==4.10.0

celery==4.4.7
tornado>=4.2.0, <6.0.0
flower==0.9.7

# JWT Auth
djangorestframework-simplejwt

typed-ast==1.4.1
